+++ bindings/nostr-ffi/src/types/time.rs
@@ -39,7 +39,7 @@ impl Timestamp {
     #[uniffi::constructor]
     pub fn from_secs(secs: u64) -> Self {
         Self {
-            inner: nostr::Timestamp::from(secs),
+            inner: nostr::Timestamp::from_secs(secs),
         }
     }